backend/requirements.txt
langchain==0.3.25
langchain_community==0.3.24
langchain_huggingface==0.2.0
numpy==2.2.6
pandas==2.2.3
protobuf==6.31.1
Requests==2.32.3
google-generativeai==0.8.5
pipreqs==0.5.0
reportlab==4.4.1
